Fix circular dep in Profile.tsx
